mystic_assistant/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
openai==0.28.0
python-dotenv==1.0.0
vosk==0.3.45
python-multipart==0.0.6
jinja2==3.1.2 